+++ bin/main.ml
@@ -250,6 +250,17 @@ let external_lib_deps =
       Arg.(non_empty
            & pos_all string []
            & Arg.info [] ~docv:"TARGET")
+    and by_dir =
+      Arg.(value
+           & flag
+           & info ["unstable-by-dir"]
+               ~doc:{|Print dependencies per directory
+                      (this feature is currently unstable)|})
+    and sexp =
+      Arg.(value
+           & flag
+           & info ["sexp"]
+               ~doc:{|Produce a s-expression output|})
     in
     Common.set_common common ~targets:[];
     let log = Log.create common in
@@ -259,21 +270,28 @@ let external_lib_deps =
     in
     let targets = Target.resolve_targets_exn ~log common setup targets in
     let request = Target.request setup targets in
+    let lib_deps = Build_system.all_lib_deps setup.build_system ~request in
     let failure =
-      String.Map.foldi ~init:false
-        (Build_system.all_lib_deps_by_context setup.build_system ~request)
-        ~f:(fun context_name lib_deps acc ->
+      String.Map.foldi lib_deps ~init:false
+        ~f:(fun context_name lib_deps_by_dir acc ->
+          let lib_deps =
+            Path.Map.values lib_deps_by_dir
+            |> List.fold_left ~init:Lib_name.Map.empty ~f:Lib_deps_info.merge
+          in
           let internals =
             Super_context.internal_lib_names
               (match String.Map.find setup.Main.scontexts context_name with
                | None -> assert false
                | Some x -> x)
           in
+          let is_external name _kind = not (Lib_name.Set.mem internals name) in
           let externals =
-            Lib_name.Map.filteri lib_deps ~f:(fun name _ ->
-              not (Lib_name.Set.mem internals name))
+            Lib_name.Map.filteri lib_deps ~f:is_external
           in
           if only_missing then begin
+            if by_dir || sexp then
+              die "@{<error>Error@}: --only-missing cannot be used with \
+                   --unstable-by-dir or --sexp";
             let context =
               List.find_exn setup.contexts ~f:(fun c -> c.name = context_name)
             in
@@ -312,7 +330,26 @@ let external_lib_deps =
                  |> String.concat ~sep:" ");
               true
             end
+          end else if sexp then begin
+            if not by_dir then
+              die "@{<error>Error@}: --sexp requires --unstable-by-dir";
+            let lib_deps_by_dir =
+              lib_deps_by_dir
+              |> Path.Map.map ~f:(Lib_name.Map.filteri ~f:is_external)
+              |> Path.Map.filter ~f:(fun m -> not (Lib_name.Map.is_empty m))
+            in
+            let sexp =
+              Map.to_sexp
+                Path.Map.to_list
+                (fun p -> Atom (Path.to_string p))
+                Lib_deps_info.to_sexp
+                lib_deps_by_dir
+            in
+            Format.printf "%a@." Sexp.pp (List [Atom context_name; sexp]);
+            acc
           end else begin
+            if by_dir then
+              die "@{<error>Error@}: --unstable-by-dir cannot be used without --sexp";
             Printf.printf
               "These are the external library dependencies in the %s context:\n\
                %s\n%!"