+++ src/lib_deps_info.mli
@@ -18,3 +18,4 @@ end
 type t = Kind.t Lib_name.Map.t
 
 val merge : t -> t -> t
+val to_sexp : t -> Sexp.t